+++ crates/amaru-consensus/src/effects/mod.rs
@@ -19,7 +19,7 @@ mod store_effects;
 
 pub use amaru_protocols::metrics_effects::{Metrics, MetricsOps, RecordMetricsEffect, ResourceMeter};
 pub use ledger_effects::{
-    Ledger, LedgerOps, RegisteredRelaySocketAddrsEffect, ResourceBlockValidation, ResourceConsensusParameters,
+    Ledger, LedgerOps, RegisteredRelayCandidatesEffect, ResourceBlockValidation, ResourceConsensusParameters,
     ResourceEraHistory, ResourceHasStakePools, ResourcePoolSummaries, ResourceTxValidation, SwitchToForkEffect,
     TipEffect, ValidateBlockEffect, ValidateHeaderEffect, ValidateTxEffect, VolatileTipEffect,
 };